requirements.txt
numpy==1.13.1
python-twitter==3.3
matplotlib==2.0.2
scipy==0.19.1
google-cloud==0.26.1